PA-669
Fixed.

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/detail/ActionBarTopView.java
@@ -401,10 +401,10 @@ public class ActionBarTopView extends LinearLayout implements WorkOrderRenderer 
 
     private final RunningLateDialog.OnSendListener _runningLateDialog_onSend = new RunningLateDialog.OnSendListener() {
         @Override
-        public void onSend(long workOrderId) {
+        public void onSend(int workOrderId) {
             if (_workOrder.getId() == workOrderId)
                 WorkOrderTracker.onActionButtonEvent(App.get(), WorkOrderTracker.ActionButton.RUNNING_LATE,
-                        WorkOrderTracker.Action.RUNNING_LATE, (int) workOrderId);
+                        WorkOrderTracker.Action.RUNNING_LATE, workOrderId);
         }
     };
